requirements.txt
discord.py
pillow
requests
python-dotenv
Flask

